Added: KERNEL_REDUCE - less than 128MB used space, Added: REDUCE_ parameters

+++ bootstrap.d/50-firstboot.sh
@@ -9,7 +9,7 @@
 cat files/firstboot/10-begin.sh > "$R/etc/rc.firstboot"
 
 # Ensure openssh server host keys are regenerated on first boot
-if [ "$ENABLE_SSHD" = true ] ; then
+if [ "$ENABLE_SSHD" = true ] && [ "$ENABLE_REDUCE" = false ]; then
   cat files/firstboot/21-generate-ssh-keys.sh >> "$R/etc/rc.firstboot"
   rm -f "$R/etc/ssh/ssh_host_*"
 fi
